Solved the foreign key constraint issue and updated fields for products returned from the product API

+++ src/main/java/com/example/orderapi/dto/OrderDTO.java
@@ -11,7 +11,7 @@ public class OrderDTO {
     private Long id;
     private String guestEmail;
     private String couponCode;
-    private Integer couponID;
+    private Long couponID;
     private BigDecimal amount;
     private Date createdAt;
     private List<OrderItemDTO> orderItems;
